requirements.txt
mysql-connector-python==9.0.0
click==8.1.7
rich==13.7.1
tabulate==0.9.0
